# scripts/sdk/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nocodb",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
    "/api/v1/db/auth/user/signup": {
      "post": {
        "summary": "Signup",
        "operationId": "auth-signup",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "Successful registration response": {
                    "value": {
                      "token": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "Invalid email": {
                    "value": {
                      "msg": "Invalid email"
                    }
                  },
                  "Invalid invite url": {
                    "value": {
                      "msg": "Invalid invite url"
                    }
                  },
                  "Expired invite url": {
                    "value": {
                      "msg": "Expired invite url, Please contact super admin to get a new invite url"
                    }
                  },
                  "User already exist": {
                    "value": {
                      "msg": "User already exist"
                    }
                  },
                  "Invite only signup": {
                    "value": {
                      "msg": "Not allowed to signup, contact super admin"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/Signup"
        },
        "description": "Create a new user with provided email and password and first user is marked as super admin. "
      }
    },
    "/api/v1/db/auth/user/signin": {
      "post": {
        "summary": "Signin",
        "operationId": "auth-signin",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              },
              "examples": {
                "example-1": {
                  "value": {
                    "email": "user@nocodb.com",
                    "password": "Password"
                  }
                }
              }
            }
          }
        },
        "description": "Authenticate existing user with their email and password. Successful login will return a JWT access-token. "
      },
      "parameters": []
    },
    "/api/v1/db/auth/user/me": {
      "parameters": [],
      "get": {
        "summary": "User Info",
        "operationId": "auth-me",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfo"
                },
                "examples": {
                  "example-1": {
                    "value": {
                      "id": "string",
                      "email": "string",
                      "email_verified": "string",
                      "firstname": "string",
                      "lastname": "string",
                      "roles": {
                        "editor": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ],
        "description": "Returns authenticated user info",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "project_id",
            "description": "Pass project id to get project specific roles along with user info"
          }
        ]
      }
    },
    "/api/v1/db/auth/password/forgot": {
      "post": {
        "summary": "Password Forgot",
        "operationId": "auth-password-forgot",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "description": "Emails user with a reset url.",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "description": "Pass registered user email id in request body"
        }
      },
      "parameters": []
    },
...
    "/api/v1/db/auth/token/validate/{token}": {
      "post": {
        "summary": "Reset Token Verify",
        "operationId": "auth-password-reset-token-validate",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Validtae password reset url token.",
        "tags": [
          "Auth"
        ]
      },
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "token",
          "in": "path",
          "required": true
        }
      ]
    },
    "/api/v1/db/auth/email/validate/{token}": {
      "post": {
        "summary": "Verify Email",
        "operationId": "auth-email-validate",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Api for verifying email where token need to be passed which is shared to user email.",
        "tags": [
          "Auth"
        ]
      },
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "token",
          "in": "path",
          "required": true
        }
      ]
    },
    "/api/v1/db/auth/password/reset/{token}": {
      "post": {
        "summary": "Password Reset",
        "operationId": "auth-password-reset",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Update user password to new by using reset token.",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "new_password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "token",
          "in": "path",
          "required": true
        }
      ]
    },
...
  },
  "components": {
    "schemas": {
...
      "UserInfo": {
        "title": "UserInfo",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "email_verified": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "roles": {}
        }
      }
    },
    "requestBodies": {
      "Signup": {
        "content": {
          "application/json": {
            "schema": {
              "description": "",
              "type": "object",
              "x-examples": {
                "example-1": {
                  "email": true
                }
              },
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          },
          "application/xml": {
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "xc-auth": {
        "name": "API Key",
        "type": "apiKey",
        "in": "header",
        "description": "JWT access token"
      },
      "xc-shared-base-id": {
        "name": "API Key",
        "type": "apiKey",
        "in": "header",
        "description": "Shared base uuid"
      }
    }
  }
}